chore: remove TODO comment
Theoretically, there is a scenario when the goroutine executing the callback in `WithTimeout()` is leaked. It would happen if the callback doesn't terminate before the timeout, and the context is cancelled. To prevent leaking the goroutine, the callback function should monitor the context and return when it's cancelled. I might be wrong, and I only skimmed through the code, but it seems like this is not the case. However, this is not critical since the error from canceling the context on timeout is lifted all the way up anyway, and the application exits with the respective exit code.